requirements.txt
pywinpty>=2.0
websockets>=12.0
pyyaml>=6.0
pygame-ce>=2.5.0
